Compile on install
Compilation was happening during a prepublish, but the dist directory
was ignored. This meant that a consumer of this package would have to
enter their node_modules directory, make the dist directory, and then
run `npm run prepublish`. There are a few solutions to this problem:

1) Don't use any ES2015 features
2) Track the dist directory
3) Generate the build artefact on install

I've chosen to use solution  3 here so that build artefacts are left out
of the repository.